# crypto/tls: reuse the tls Conn's outBuf to reduce memory cost on large number of tls connections