benchstat old.txt new.txt

name                  old time/op    new time/op    delta
ConnOutbufWithPool-8    1.95µs ± 1%    0.62µs ± 1%  -68.16%  (p=0.000 n=9+10)

name                  old alloc/op   new alloc/op   delta
ConnOutbufWithPool-8    10.6kB ± 0%     1.3kB ± 0%  -87.79%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
ConnOutbufWithPool-8      7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=10+10)

Fixes #42035