@cch123 2020-10-19
```
~/g/s/m/m/p/m/c/tls ❯❯❯ cat new.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	 2063755	       582 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2012469	       573 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2100492	       584 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2098694	       597 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 1859014	       631 ns/op	    1285 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2007234	       604 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 1957771	       682 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2039295	       599 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2059744	       587 ns/op	    1286 B/op	       2 allocs/op
BenchmarkConnOutbufWithPool-8   	 2106589	       581 ns/op	    1286 B/op	       2 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	18.754s
```
```
~/g/s/m/m/p/m/c/tls ❯❯❯ cat old.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	  520760	      1924 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  640018	      2018 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  598604	      1979 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  595036	      1974 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  619786	      2019 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  616986	      2021 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  582493	      1999 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  628933	      1903 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  631393	      1927 ns/op	   10629 B/op	       6 allocs/op
BenchmarkConnOutbufWithPool-8   	  590546	      2014 ns/op	   10629 B/op	       6 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	13.986s
```
```
~/g/s/m/m/p/m/c/tls ❯❯❯ benchstat old.txt new.txt
name                  old time/op    new time/op    delta
ConnOutbufWithPool-8    1.98µs ± 4%    0.59µs ± 6%  -70.01%  (p=0.000 n=10+9)

name                  old alloc/op   new alloc/op   delta
ConnOutbufWithPool-8    10.6kB ± 0%     1.3kB ± 0%  -87.90%  (p=0.000 n=10+9)

name                  old allocs/op  new allocs/op  delta
ConnOutbufWithPool-8      6.00 ± 0%      2.00 ± 0%  -66.67%  (p=0.000 n=10+10)

```