@cch123 2020-10-19
```
~/g/s/m/m/p/m/c/tls ❯❯❯ go test -run=none  -bench=ConnOutbufWithPool -benchmem -count=10 | tee old.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	  538656	      1977 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  606666	      1960 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  603496	      1956 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  604824	      1941 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  608166	      1967 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  526492	      2057 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  576446	      1959 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  608270	      1937 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  591682	      1940 ns/op	   10645 B/op	       7 allocs/op
BenchmarkConnOutbufWithPool-8   	  583629	      1953 ns/op	   10645 B/op	       7 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	11.817s
```
```
~/g/s/m/m/p/m/c/tls ❯❯❯ go test -run=none  -bench=ConnOutbufWithPool -benchmem -count=10 | tee new.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	 1945142	       631 ns/op	    1299 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1918827	       616 ns/op	    1299 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1923861	       624 ns/op	    1300 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1909788	       621 ns/op	    1300 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1854442	       618 ns/op	    1299 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1933804	       616 ns/op	    1300 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1926312	       628 ns/op	    1299 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1925890	       620 ns/op	    1300 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1926680	       626 ns/op	    1300 B/op	       3 allocs/op
BenchmarkConnOutbufWithPool-8   	 1926465	       623 ns/op	    1300 B/op	       3 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	18.625s
```
```
~/g/s/m/m/p/m/c/tls ❯❯❯ benchstat old.txt new.txt
name                  old time/op    new time/op    delta
ConnOutbufWithPool-8    1.95µs ± 1%    0.62µs ± 1%  -68.16%  (p=0.000 n=9+10)

name                  old alloc/op   new alloc/op   delta
ConnOutbufWithPool-8    10.6kB ± 0%     1.3kB ± 0%  -87.79%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
ConnOutbufWithPool-8      7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=10+10)
```