@cch123 2020-10-19
in our scene, this can save a lot of memory, here are data for 50,000 TLS connections, 5000 QPS:

![image](https://user-images.githubusercontent.com/384546/96449081-e42b8480-1246-11eb-922f-1a6d990dbdb4.png)