@cch123 2020-10-23
Detailed Benchmark:

## Max QPS test

GOMAXPROCS=16. / wrk -c 1024 -d 10s -t64 https://localhost:4443

Original TLS:

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 249252.58 | 247981.13 | 252072.94|  246838.65|  248069.50|  250279.61|  248893.85 |  248653.03 | 

Use sync.Pool:

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 233128.73 | 236434.06 | 241020.74 | 237904.41 | 237683.17|  240238.59 | 241162.77 |  238880.61|

GOMAXPROCS=8. / wrk -c 1024 -d 10s -t64 https://localhost:4443

Original TLS

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 137442.72| 135192.82 |  137283.12| 138932.01| 136880.19 |  137972.56|  139443.37 |  140283.15|


Use sync.Pool:
|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 134797.43 |132957.91| 132796.81| 134416.20| 136944.84| 135315.26| 134220.55| 133366.97|

https server:
```go
package main

import (
	"fmt"
	"net/http"
	_ "net/http/pprof"
)

var content = make([]byte, 16000)

func sayhello(wr http.ResponseWriter, r *http.Request) {
	wr.Header()["Content-Length"] = []string{fmt.Sprint(len(content))}
	wr.Header()["Content-Type"] = []string{"application/json"}
	wr.Write(content)
}

func main() {
	go func() {
		http.ListenAndServe(":3333", nil)
	}()
	http.HandleFunc("/", sayhello)

	err := http.ListenAndServeTLS(":4443", "server.crt", "server.key", nil)
	if err != nil {
		fmt.Println(err)
	}
}
```