@cch123 2020-10-23
## Short connections max QPS test

GOMAXPROCS=8
./wrk -H "Connection: Close" -c 2048 -d 10s -t64 https://localhost:4443

Original TLS:

|seq|1|2|3|4|5|
|--|--|--|--|--|--|
|QPS|9860.85|9881.71|9957.22|9819.28|9884.17|

Use sync.Pool

|seq|1|2|3|4|5|
|--|--|--|--|--|--|
|QPS|10962.59|10834.51|11084.84|10844.41|10691.28|

GOMAXPROCS=16
$sudo ./wrk -H "Connection: Close" -c 4096 -d 10s -t64 https://localhost:4443

Original TLS:

|seq|1|2|3|4|5|
|--|--|--|--|--|--|
|QPS|16103.86|16132.89|16267.94|15949.50|15352.71|

Use sync.Pool
|seq|1|2|3|4|5|
|--|--|--|--|--|--|
|QPS|16914.97|17324.14|17379.78|17070.90|17500.04|